Add docker-compose to run etcd and tests
On many systems running etcd is a burden of a requirement, in this setup
you can run etcd in a container and also if required run your tests in a
container with the 'etcdctl' binary.

The tests support an optional etcd enpoint environment variable
'ETCD_ENDPOINT'. If this is supplied then this is is the endpoint that
is connected to.

You can run the full test suite with `docker-compose up` which will
leave the system running. To run etcd up in the background, use `docker-compose up -d
etcd`, this will expose etcd on port 2379 on your localhost.

Once you have a etcd cluster to use the tests can either be ran locally
using the standard `tox` command and using the env var `ETCD_ENDPOINT`
or you can run the test in the docker container if you do not have the
`etcdctl` binary available. This can be done with `docker-compose up
python`.

If a code change is made the 'python' container will need to be rebuilt,
this should be instant process if the `tox.ini` and
`dev-requirements.txt` have not changed. Run `docker-compose build` to
rebuild the container with the new code.