Handle grpc exceptions during watch
The watch events iterator never exits if a grpc error occurs (such as a
connection error) and the end user was waitng for events even though
the connection had been dropped. Now, if an exception happens, it is
propageted to the events' iterators through the event callback.

Also now, the client's timeout value is used during a watch creation.
If the timeout expires (indicating a grpc issue) and the watch has not
been established, an exception is raised.

Fixes #88