Allow defining multiple endpoints to connect to
Since etcd is a distributed storage, it makes sense for a client library
to be able to connect to all the nodes in the cluster, and to manage
their current status without needing to be reinstantiated.

This patch adds the ability for the client to be aware of the status of
all nodes in the cluster, and - in case of failure - to mark each of
them as temporarily failed.

Individual requests will continue to raise an exception in case of a
failed connection at this point, but it would be easy to allow retrying
if it seems a good idea.

What has been done in practice:

* Added a 'Endpoint' class, a simple FSM to abstract a remote server.
* Refactored the initialization of the library, adding a switch to allow
  failing over to another server in case of need.
* Refactored self.channel to be a function returning a grpc channel to
  the first non-failed node, and all the Etcd3Client.*stub properties
  accordingly.
* Refactored the _handle_error wrapper, made it part of the class, and
  split it between the version for generators and the one for normal
  returners for code simplicity