Adapt the failover to be compatible with asyncio
Remove code from the channel property setter. The setter property code was too complex to be in a
property. It was then not possible to adapt it for asyncio.
The choose of the new endpoint and the clear_old_stubs was perfom in the channel setter.
I moved it to the _manage_grpc_errors to avoid to replace the channel setter to a new method.